# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/appointments": {
            "get": {
                "tags": [
                    "Appointment"
                ],
                "summary": "Get a listing of the Appointments.",
                "description": "Get all Appointments",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Appointment"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Appointment"
                ],
                "summary": "Store a newly created Appointment in storage",
                "description": "Store Appointment",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Appointment that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Appointment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Appointment"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/appointments/{id}": {
            "get": {
                "tags": [
                    "Appointment"
                ],
                "summary": "Display the specified Appointment",
                "description": "Get Appointment",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Appointment",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Appointment"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Appointment"
                ],
                "summary": "Update the specified Appointment in storage",
                "description": "Update Appointment",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Appointment",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Appointment that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Appointment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Appointment"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Appointment"
                ],
                "summary": "Remove the specified Appointment from storage",
                "description": "Delete Appointment",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Appointment",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Appointment": {
            "required": [
                "shop_id",
                "client_id",
                "employee_id",
                "service_id",
                "date",
                "time"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "shop_id": {
                    "description": "shop_id",
                    "type": "integer",
                    "format": "int32"
                },
                "client_id": {
                    "description": "client_id",
                    "type": "integer",
                    "format": "int32"
                },
                "employee_id": {
                    "description": "employee_id",
                    "type": "integer",
                    "format": "int32"
                },
                "service_id": {
                    "description": "service_id",
                    "type": "integer",
                    "format": "int32"
                },
                "date": {
                    "description": "date",
                    "type": "string",
                    "format": "date"
                },
                "time": {
                    "description": "time",
                    "type": "string"
                },
                "comments": {
                    "description": "comments",
                    "type": "string"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
    }
}